Extract performed on class WaitingQueueManager -> WaitingQueueManager and WaitingQueueManagerReservation
